incorporate NWB loading into load_poses module
